Add some audit logging to more services and better security boundary checking
