+++ application/backend/src/business/services/release-participation-service.ts
@@ -79,106 +79,78 @@ export class ReleaseParticipationService extends ReleaseBaseService {
       releaseKey
     );
 
-    const now = new Date();
-    const newAuditEventId = await this.auditLogService.startReleaseAuditEvent(
-      this.edgeDbClient,
+    return await this.auditLogService.transactionalUpdateInReleaseAuditPattern(
       user,
       releaseKey,
-      "E",
       "Add Participant",
-      now
-    );
-
-    try {
-      // right to add participants must be limited to DataOwners and PIs
-      if (userRole !== "DataOwner" && userRole !== "PI") {
-        throw new ReleaseParticipationPermissionError(releaseKey);
-      }
-
-      const userOrPotentialUserUuid = await this.edgeDbClient.transaction(
-        async (tx) => {
-          // we have three scenarios to handle
-          // (1) the user exists in Elsa and we just need to link them to the release
-          // (2) the user has not logged in but does exist already as a PotentialUser - we need to link them to the release
-          // (3) the user is completely new and we need to create and link as a PotentialUser
-
-          // (1)
-          const dbUser = await singleUserByEmailQuery.run(tx, {
-            email: newUserEmail,
-          });
-
-          if (dbUser) {
-            await releaseParticipantAddUser(tx, {
-              userUuid: dbUser.id,
-              releaseKey: releaseKey,
-              role: newUserRole,
-            });
-            return dbUser.id;
-          }
-
-          // (2)
-          const potentialDbUser = await singlePotentialUserByEmailQuery.run(
-            tx,
-            {
-              email: newUserEmail,
-            }
-          );
-
-          if (potentialDbUser) {
-            await releaseParticipantAddPotentialUser(tx, {
-              potentialUserUuid: potentialDbUser.id,
-              releaseKey: releaseKey,
-              role: newUserRole,
-            });
-            return potentialDbUser.id;
-          }
-
-          // (3)
-          const newPotential = await e
-            .insert(e.permission.PotentialUser, {
-              displayName: newUserEmail,
-              email: newUserEmail,
-              futureReleaseParticipant: e.select(e.release.Release, (r) => ({
-                filter: e.op(releaseKey, "=", r.releaseKey),
-                "@role": e.str(newUserRole),
-              })),
-            })
-            .assert_single()
-            .run(tx);
-
-          if (newPotential) return newPotential.id;
-          else throw new Error("New Potential User failed to add");
+      async () => {
+        // right to add participants must be limited to DataOwners and PIs
+        if (userRole !== "DataOwner" && userRole !== "PI") {
+          throw new ReleaseParticipationPermissionError(releaseKey);
         }
-      );
-
-      await this.auditLogService.completeReleaseAuditEvent(
-        this.edgeDbClient,
-        newAuditEventId,
-        0,
-        now,
-        new Date(),
-        {
+      },
+      async (tx, a) => {
+        // a datastructure we pass to the next stage AND which we put into the Audit log details
+        const auditLogDetail = (id: string) => ({
           email: newUserEmail,
           role: newUserRole,
-          affectedUserOrPotentialUser: userOrPotentialUserUuid,
+          affectedUserOrPotentialUser: id,
+        });
+
+        // we have three scenarios to handle
+        // (1) the user exists in Elsa and we just need to link them to the release
+        // (2) the user has not logged in but does exist already as a PotentialUser - we need to link them to the release
+        // (3) the user is completely new and we need to create and link as a PotentialUser
+
+        // (1)
+        const dbUser = await singleUserByEmailQuery.run(tx, {
+          email: newUserEmail,
+        });
+
+        if (dbUser) {
+          await releaseParticipantAddUser(tx, {
+            userUuid: dbUser.id,
+            releaseKey: releaseKey,
+            role: newUserRole,
+          });
+          return auditLogDetail(dbUser.id);
         }
-      );
 
-      return userOrPotentialUserUuid;
-    } catch (e) {
-      const errorString = e instanceof Error ? e.message : String(e);
+        // (2)
+        const potentialDbUser = await singlePotentialUserByEmailQuery.run(tx, {
+          email: newUserEmail,
+        });
 
-      await this.auditLogService.completeReleaseAuditEvent(
-        this.edgeDbClient,
-        newAuditEventId,
-        8,
-        now,
-        new Date(),
-        { error: errorString }
-      );
+        if (potentialDbUser) {
+          await releaseParticipantAddPotentialUser(tx, {
+            potentialUserUuid: potentialDbUser.id,
+            releaseKey: releaseKey,
+            role: newUserRole,
+          });
+          return auditLogDetail(potentialDbUser.id);
+        }
 
-      throw e;
-    }
+        // (3)
+        const newPotential = await e
+          .insert(e.permission.PotentialUser, {
+            displayName: newUserEmail,
+            email: newUserEmail,
+            futureReleaseParticipant: e.select(e.release.Release, (r) => ({
+              filter: e.op(releaseKey, "=", r.releaseKey),
+              "@role": e.str(newUserRole),
+            })),
+          })
+          .assert_single()
+          .run(tx);
+
+        if (newPotential) return auditLogDetail(newPotential.id);
+
+        throw new Error("New Potential User failed to add");
+      },
+      async (a) => {
+        return a.affectedUserOrPotentialUser;
+      }
+    );
   }
 
   public async removeParticipant(
@@ -191,72 +163,43 @@ export class ReleaseParticipationService extends ReleaseBaseService {
       releaseKey
     );
 
-    const now = new Date();
-    const newAuditEventId = await this.auditLogService.startReleaseAuditEvent(
-      this.edgeDbClient,
+    await this.auditLogService.transactionalUpdateInReleaseAuditPattern(
       user,
       releaseKey,
-      "E",
       "Remove Participant",
-      now
-    );
+      async () => {
+        // deleting participants must be limited to DataOwners and PIs
+        if (userRole !== "DataOwner" && userRole !== "PI") {
+          throw new ReleaseParticipationPermissionError(releaseKey);
+        }
+      },
+      async (tx, a) => {
+        // NOTE: we use the normal EdgeDb semantics to say that doing a "-=" where the unliked item IS NOT PRESENT
+        // IS NOT AN ERROR i.e. deleting someone from a release who is not in the release
+        // WE COULD DECIDE TO DO THIS DIFFERENTLY (?)
 
-    try {
-      // deleting participants must be limited to DataOwners and PIs
-      if (userRole !== "DataOwner" && userRole !== "PI") {
-        throw new ReleaseParticipationPermissionError(releaseKey);
-      }
+        // we are actually pretty loose here on checking the participant and release existence.. but I can't
+        // think of any way it affects things
+        // if the participantUuid is invalid - then nothing happens
+        // if the participantUuid is not actually in *this* release - then nothing happens
+        // the releaseUuid *is* tested so we know the instigator exists and has some permissions
 
-      const actuallyRemoved = await this.edgeDbClient.transaction(
-        async (tx) => {
-          // NOTE: we use the normal EdgeDb semantics to say that doing a "-=" where the unliked item IS NOT PRESENT
-          // IS NOT AN ERROR i.e. deleting someone from a release who is not in the release
-          // WE COULD DECIDE TO DO THIS DIFFERENTLY (?)
-
-          // we are actually pretty loose here on checking the participant and release existence.. but I can't
-          // think of any way it affects things
-          // if the participantUuid is invalid - then nothing happens
-          // if the participantUuid is not actually in *this* release - then nothing happens
-          // the releaseUuid *is* tested so we know the instigator exists and has some permissions
-
-          const userRemoved = await releaseParticipantRemoveUser(tx, {
-            userUuid: participantUuid,
+        const userRemoved = await releaseParticipantRemoveUser(tx, {
+          userUuid: participantUuid,
+          releaseKey: releaseKey,
+        });
+        const potentialUserRemoved =
+          await releaseParticipantRemovePotentialUser(tx, {
+            potentialUserUuid: participantUuid,
             releaseKey: releaseKey,
           });
-          const potentialUserRemoved =
-            await releaseParticipantRemovePotentialUser(tx, {
-              potentialUserUuid: participantUuid,
-              releaseKey: releaseKey,
-            });
 
-          // altering the participation of a release counts as a touch
-          await touchRelease.run(tx, { releaseKey: releaseKey });
-
-          // return true if we actually removed someone, false otherwise
-          return userRemoved || potentialUserRemoved;
-        }
-      );
-
-      await this.auditLogService.completeReleaseAuditEvent(
-        this.edgeDbClient,
-        newAuditEventId,
-        actuallyRemoved ? OUTCOME_SUCCESS : OUTCOME_MINOR_FAILURE,
-        now,
-        new Date()
-      );
-    } catch (e) {
-      const errorString = e instanceof Error ? e.message : String(e);
-
-      await this.auditLogService.completeReleaseAuditEvent(
-        this.edgeDbClient,
-        newAuditEventId,
-        OUTCOME_SERIOUS_FAILURE,
-        now,
-        new Date(),
-        { error: errorString }
-      );
-
-      throw e;
-    }
+        return {
+          actualUserRemoved:
+            userRemoved?.id ?? potentialUserRemoved?.id ?? "none",
+        };
+      },
+      async (a) => {}
+    );
   }
 }
