+++ application/backend/tests/service-tests/releases.cases.test.ts
@@ -25,9 +25,10 @@ import {
   JETSONS_CASE,
   JUDY_SPECIMEN,
 } from "../../src/test-data/insert-test-data-10f-jetsons";
+import { ReleaseSelectionService } from "../../src/business/services/release-selection-service";
 
 let edgeDbClient: Client;
-let releaseService: ReleaseService;
+let releaseSelectionService: ReleaseSelectionService;
 let testReleaseKey: string;
 
 let allowedDataOwnerUser: AuthenticatedUser;
@@ -41,7 +42,7 @@ beforeAll(async () => {
   const testContainer = await registerTypes();
 
   edgeDbClient = testContainer.resolve("Database");
-  releaseService = testContainer.resolve(ReleaseService);
+  releaseSelectionService = testContainer.resolve(ReleaseSelectionService);
 });
 
 beforeEach(async () => {
@@ -53,7 +54,7 @@ beforeEach(async () => {
  *
  */
 it("get all case level information from a release as a data owner", async () => {
-  const pagedResult = await releaseService.getCases(
+  const pagedResult = await releaseSelectionService.getCases(
     allowedDataOwnerUser,
     testReleaseKey,
     DEFAULT_LIMIT,
@@ -93,7 +94,7 @@ it("get all case level information from a release as a data owner", async () => 
  *
  */
 it("get limited case level information from a release as a PI", async () => {
-  const pagedResult = await releaseService.getCases(
+  const pagedResult = await releaseSelectionService.getCases(
     allowedPiUser,
     testReleaseKey,
     DEFAULT_LIMIT,
@@ -135,7 +136,7 @@ it("get limited case level information from a release as a PI", async () => {
  *
  */
 it("get patient/specimen level data fields", async () => {
-  const pagedResult = await releaseService.getCases(
+  const pagedResult = await releaseSelectionService.getCases(
     allowedDataOwnerUser,
     testReleaseKey,
     DEFAULT_LIMIT,
@@ -161,7 +162,7 @@ it("get patient/specimen level data fields", async () => {
  */
 it("node status changes as leaves are selected and unselected", async () => {
   {
-    const initialResult = await releaseService.getCases(
+    const initialResult = await releaseSelectionService.getCases(
       allowedDataOwnerUser,
       testReleaseKey,
       DEFAULT_LIMIT,
@@ -197,7 +198,7 @@ it("node status changes as leaves are selected and unselected", async () => {
     );
   }
 
-  await releaseService.setSelected(
+  await releaseSelectionService.setSelected(
     allowedDataOwnerUser,
     testReleaseKey,
     await findDatabaseSpecimenIds(edgeDbClient, [
@@ -208,7 +209,7 @@ it("node status changes as leaves are selected and unselected", async () => {
   );
 
   {
-    const afterSetResult = await releaseService.getCases(
+    const afterSetResult = await releaseSelectionService.getCases(
       allowedDataOwnerUser,
       testReleaseKey,
       DEFAULT_LIMIT,
@@ -246,14 +247,14 @@ it("node status changes as leaves are selected and unselected", async () => {
     );
   }
 
-  await releaseService.setUnselected(
+  await releaseSelectionService.setUnselected(
     allowedDataOwnerUser,
     testReleaseKey,
     await findDatabaseSpecimenIds(edgeDbClient, [BART_SPECIMEN, HOMER_SPECIMEN])
   );
 
   {
-    const afterUnsetResult = await releaseService.getCases(
+    const afterUnsetResult = await releaseSelectionService.getCases(
       allowedDataOwnerUser,
       testReleaseKey,
       DEFAULT_LIMIT,
@@ -282,7 +283,7 @@ it("node status changes as leaves are selected and unselected", async () => {
 
 it("(un-)selects all when setSelectedStatus is passed an empty list", async () => {
   const allSpecimens_ = async (): Promise<ReleaseSpecimenType[]> => {
-    const result = await releaseService.getCases(
+    const result = await releaseSelectionService.getCases(
       allowedDataOwnerUser,
       testReleaseKey,
       DEFAULT_LIMIT,
@@ -299,12 +300,20 @@ it("(un-)selects all when setSelectedStatus is passed an empty list", async () =
     return specimens;
   };
 
-  await releaseService.setSelected(allowedDataOwnerUser, testReleaseKey, []);
+  await releaseSelectionService.setSelected(
+    allowedDataOwnerUser,
+    testReleaseKey,
+    []
+  );
   expect(
     (await allSpecimens_()).every((s) => s.nodeStatus === "selected")
   ).toBe(true);
 
-  await releaseService.setUnselected(allowedDataOwnerUser, testReleaseKey, []);
+  await releaseSelectionService.setUnselected(
+    allowedDataOwnerUser,
+    testReleaseKey,
+    []
+  );
   expect(
     (await allSpecimens_()).every((s) => s.nodeStatus === "unselected")
   ).toBe(true);
@@ -312,7 +321,7 @@ it("(un-)selects all when setSelectedStatus is passed an empty list", async () =
 
 it("pass in specimen ids that are not valid", async () => {
   await expect(async () => {
-    await releaseService.setSelected(
+    await releaseSelectionService.setSelected(
       allowedPiUser,
       testReleaseKey,
       // whilst this looks vaguely like a edgedb id it will never match
@@ -342,7 +351,7 @@ it("pass in specimen ids that are not valid", async () => {
   do {
     page += 1;
 
-    result = await releaseService.getCases(
+    result = await releaseSelectionService.getCases(
       allowedDataOwnerUser,
       testReleaseKey,
       limit,
@@ -363,7 +372,7 @@ it("pass in specimen ids that are not valid", async () => {
 }); */
 
 it("test identifier searching with case level match", async () => {
-  const result = await releaseService.getCases(
+  const result = await releaseSelectionService.getCases(
     allowedDataOwnerUser,
     testReleaseKey,
     DEFAULT_LIMIT,
@@ -382,7 +391,7 @@ it("test identifier searching with case level match", async () => {
 });
 
 it("test identifier searching with patient level match", async () => {
-  const result = await releaseService.getCases(
+  const result = await releaseSelectionService.getCases(
     allowedDataOwnerUser,
     testReleaseKey,
     DEFAULT_LIMIT,
@@ -399,7 +408,7 @@ it("test identifier searching with patient level match", async () => {
 });
 
 it("test identifier searching with specimen level match", async () => {
-  const result = await releaseService.getCases(
+  const result = await releaseSelectionService.getCases(
     allowedDataOwnerUser,
     testReleaseKey,
     DEFAULT_LIMIT,
@@ -416,7 +425,7 @@ it("test identifier searching with specimen level match", async () => {
 });
 
 it("test identifier searching with specimen level partial match (not supported)", async () => {
-  const result = await releaseService.getCases(
+  const result = await releaseSelectionService.getCases(
     allowedDataOwnerUser,
     testReleaseKey,
     DEFAULT_LIMIT,
